requirements.txt
arrow==0.15.5
bcrypt==3.1.7
binaryornot==0.4.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cookiecutter==1.7.0
cryptography==2.8
Flask==1.1.1
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jinja2-time==0.2.0
lxml==4.4.2
MarkupSafe==1.1.1
ncclient==0.6.7
paramiko==2.7.1
poyo==0.5.0
pyaml==19.12.0
pyang==2.1.1
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
six==1.14.0
urllib3==1.25.7
Werkzeug==0.16.0
whichcraft==0.6.1
xmltodict==0.12.0
